Move query package
